requirements.txt
comtypes==1.1.7
Django==2.2.3
django-authtools==1.7.0
pycryptodome==3.8.2
pypiwin32==223
pytz==2019.1
pywin32==224
sqlparse==0.3.0

# # Local installed
# six==1.12.0
# typed-ast==1.4.0
# wrapt==1.11.2
# pylint==2.3.1
# mccabe==0.6.1
# isort==4.3.21
# lazy-object-proxy==1.4.1
# astroid==2.2.5
# colorama==0.4.1